Align master

+++ models/DataObject/Concrete/Dao.php
@@ -27,6 +27,8 @@ use Pimcore\Model\DataObject\ClassDefinition\Data\QueryResourcePersistenceAwareI
 use Pimcore\Model\DataObject\ClassDefinition\Data\ResourcePersistenceAwareInterface;
 
 /**
+ * @internal
+ *
  * @property \Pimcore\Model\DataObject\Concrete $model
  */
 class Dao extends Model\DataObject\AbstractObject\Dao
@@ -143,6 +145,10 @@ class Dao extends Model\DataObject\AbstractObject\Dao
      */
     public function getData()
     {
+        if (empty($this->model->getClass())) {
+            return;
+        }
+
         $data = $this->db->fetchRow('SELECT * FROM object_store_' . $this->model->getClassId() . ' WHERE oo_id = ?', $this->model->getId());
 
         $fieldDefinitions = $this->model->getClass()->getFieldDefinitions(['object' => $this->model]);
@@ -152,8 +158,10 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     // datafield has it's own loader
                     $params = [
                         'context' => [
-                            'object' => $this->model
-                        ]
+                            'object' => $this->model,
+                        ],
+                        'owner' => $this->model,
+                        'fieldname' => $key,
                     ];
                     $value = $value->load($this->model, $params);
                     if ($value === 0 || !empty($value)) {
@@ -170,7 +178,9 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     }
                     $this->model->setValue($key, $value->getDataFromResource($multidata));
                 } else {
-                    $this->model->setValue($key, $value->getDataFromResource($data[$key], $this->model));
+                    $this->model->setValue($key, $value->getDataFromResource($data[$key], $this->model,
+                            ['owner' => $this->model,
+                                'fieldname' => $key, ]));
                 }
             }
         }
@@ -191,8 +201,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
         $db = Db::get();
 
         foreach ($fieldDefinitions as $key => $fd) {
-            if (($fd instanceof LazyLoadingSupportInterface || method_exists($fd, 'getLazyLoading'))
-                                    && $fd->getLazyLoading()) {
+            if ($fd instanceof LazyLoadingSupportInterface && $fd->getLazyLoading()) {
                 if (!$this->model->isLazyKeyLoaded($key) || $fd instanceof DataObject\ClassDefinition\Data\ReverseManyToManyObjectRelation) {
                     //this is a relation subject to lazy loading - it has not been loaded
                     $untouchable[] = $key;
@@ -200,7 +209,7 @@ class Dao extends Model\DataObject\AbstractObject\Dao
             }
 
             if (!DataObject\AbstractObject::isDirtyDetectionDisabled() && $fd->supportsDirtyDetection()) {
-                if ($this->model instanceof DataObject\DirtyIndicatorInterface && !$this->model->isFieldDirty($key)) {
+                if ($this->model instanceof Model\Element\DirtyIndicatorInterface && !$this->model->isFieldDirty($key)) {
                     if (!in_array($key, $untouchable)) {
                         $untouchable[] = $key;
                     }
@@ -235,9 +244,13 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                 $saveParams = ['isUntouchable' => in_array($fd->getName(), $untouchable),
                     'isUpdate' => $isUpdate,
                     'context' => [
-                        'containerType' => 'object'
-                    ]];
-                if ($this->model instanceof DataObject\DirtyIndicatorInterface) {
+                        'containerType' => 'object',
+                    ],
+                    'owner' => $this->model,
+                    'fieldname' => $key,
+                ]
+                ;
+                if ($this->model instanceof Model\Element\DirtyIndicatorInterface) {
                     $saveParams['newParent'] = $this->model->isFieldDirty('o_parentId');
                 }
                 $fd->save($this->model, $saveParams);
@@ -248,7 +261,8 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     $insertDataArray = $fd->getDataForResource($this->model->$getter(), $this->model,
                         [
                             'isUpdate' => $isUpdate,
-                            'owner' => $this->model
+                            'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     if (is_array($insertDataArray)) {
                         $data = array_merge($data, $insertDataArray);
@@ -257,7 +271,8 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     $insertData = $fd->getDataForResource($this->model->$getter(), $this->model,
                         [
                             'isUpdate' => $isUpdate,
-                            'owner' => $this->model
+                            'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     $data[$key] = $insertData;
                 }
@@ -295,7 +310,8 @@ class Dao extends Model\DataObject\AbstractObject\Dao
                     $insertData = $fd->getDataForQueryResource($fieldValue, $this->model,
                         [
                             'isUpdate' => $isUpdate,
-                            'owner' => $this->model
+                            'owner' => $this->model,
+                            'fieldname' => $key,
                         ]);
                     $isEmpty = $fd->isEmpty($fieldValue);
 
@@ -388,8 +404,8 @@ class Dao extends Model\DataObject\AbstractObject\Dao
     {
         $this->inheritanceHelper->doUpdate($this->model->getId(), false, [
             'inheritanceRelationContext' => [
-                'ownerType' => 'object'
-            ]
+                'ownerType' => 'object',
+            ],
         ]);
         $this->inheritanceHelper->resetFieldsToCheck();
     }
@@ -404,9 +420,11 @@ class Dao extends Model\DataObject\AbstractObject\Dao
         $this->db->delete('object_relations_' . $this->model->getClassId(), ['src_id' => $this->model->getId()]);
 
         // delete fields which have their own delete algorithm
-        foreach ($this->model->getClass()->getFieldDefinitions() as $fd) {
-            if ($fd instanceof CustomResourcePersistingInterface) {
-                $fd->delete($this->model);
+        if ($this->model->getClass()) {
+            foreach ($this->model->getClass()->getFieldDefinitions() as $fd) {
+                if ($fd instanceof CustomResourcePersistingInterface) {
+                    $fd->delete($this->model);
+                }
             }
         }
